New state for task
Added the status READY for the task. In this status the task has everything
to be submitted to the remote infrastructure but not yet submitted.
The following status, SCHEDULED, identify the condition of submitted but not running.

Additionally, the mime type has been update with a custom vendor specific mime type.